requirements.txt
# Dependências opcionais (a aplicação roda sem elas)
Pillow>=9.0.0 ; python_version >= '3.10'

